fixing compiler warnings for getCellColor()

+++ src/algorithms.c
@@ -54,7 +54,7 @@ void astar() {
         drawCell(current);
 
         // No diagonals, hence 4 neighbours. DIRS appendable to add diagonals.
-        for (int i = 0; i < sizeof(DIRS)/sizeof(DIRS[0]); i++) {
+        for (unsigned long  i = 0; i < sizeof(DIRS)/sizeof(DIRS[0]); i++) {
             // See DIRS array.
             int nx = current->x + DIRS[i][0];
             int ny = current->y + DIRS[i][1];
@@ -109,7 +109,7 @@ void BFS() {
         }
         drawCell(current);
 
-        for (int i = 0; i < sizeof(DIRS)/sizeof(DIRS[0]); i++) {
+        for (unsigned long int i = 0; i < sizeof(DIRS)/sizeof(DIRS[0]); i++) {
             // See DIRS array.
             int nx = current->x + DIRS[i][0];
             int ny = current->y + DIRS[i][1];
@@ -156,7 +156,7 @@ void DFS() {
             break;
         }
         
-        for (int i = 0; i < sizeof(DIRS)/sizeof(DIRS[0]); i++) {
+        for (unsigned long int i = 0; i < sizeof(DIRS)/sizeof(DIRS[0]); i++) {
             int nx = current->x + DIRS[i][0];
             int ny = current->y + DIRS[i][1];
